pyproject.toml
[tool.poetry]
name = "form"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.48.0"
behave = "^1.2.6"
python-dotenv = "^1.0.1"
pathlib = "^1.0.1"
allure-behave = "^2.13.5"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
